refactor(timer): extract create modal into its own component
